requirements.txt
opencv-python
pillow
qrcode[pil]
pyzbar
numpy
